+++ hamingja_dungeon/utils/exceptions.py
