+++ tests/test_hallway_designer.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 
-from hamingja_dungeon.areas.rooms.hallways.hallway_designer import HallwayDesigner
-from hamingja_dungeon.areas.rooms.room import Room
-from hamingja_dungeon.areas.vector import Vector
-from hamingja_dungeon.direction.direction import Direction
+from hamingja_dungeon.hallway_designers.hallway_designer import HallwayDesigner
+from hamingja_dungeon.dungeon_elements.room import Room
+from hamingja_dungeon.utils.vector import Vector
+from hamingja_dungeon.utils.direction import Direction
 from test_utils import (
     get_test_dungeon_area,
     print_dungeon_object,
