+++ tests/test_room.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from hamingja_dungeon.areas.rooms.room import CircleRoom
+from hamingja_dungeon.dungeon_elements.room import CircleRoom
 from test_utils import print_dungeon_object
